+++ model/command/color/ColorStaticKelvinCommand.py
@@ -5,7 +5,7 @@ from utilities.kelvin2rgb import kelvin2rgb
 
 class ColorStaticKelvinCommand(CommandBase):
     mode: Literal["color_static_kelvin"] = "color_static_kelvin"
-    kelvin: int = Field(..., ge=1000, le=12000)
+    kelvin: float = Field(..., ge=1000, le=12000)
     is_static = True
 
     def _compute(self, current_red: List[float], current_green: List[float], current_blue: List[float], targets: Iterable[int], time: float):
