+++ model/command/color/ColorStaticRgbCommand.py
@@ -4,9 +4,9 @@ from model.command.command import CommandBase
 
 class ColorStaticRgbCommand(CommandBase):
     mode: Literal["color_static_rgb"] = "color_static_rgb"
-    red: int = Field(..., ge=0, le=255)
-    green: int = Field(..., ge=0, le=255)
-    blue: int = Field(..., ge=0, le=255)
+    red: float = Field(..., ge=0, le=255)
+    green: float = Field(..., ge=0, le=255)
+    blue: float = Field(..., ge=0, le=255)
     is_static = True
 
     def _compute(self, current_red: List[float], current_green: List[float], current_blue: List[float], targets: Iterable[int], time: float):
